moving DR GUI panels to separate packacke and writing some unit tests for AnalysisUtils methods
